requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
commandlines==0.4.1
cycler==0.10.0
decorator==4.4.2
fastapi==0.65.1
filetype==1.0.7
Flask==2.0.0
h11==0.12.0
hsh==1.1.0
idna==2.10
imageio==2.9.0
itsdangerous==2.0.0
Jinja2==3.0.0
kiwisolver==1.3.1
llvmlite==0.36.0
MarkupSafe==2.0.0
matplotlib==3.4.2
networkx==2.5.1
numba==0.53.1
numpy==1.20.3
Pillow==8.2.0
pydantic==1.8.2
PyMatting==1.1.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-multipart==0.0.5
PyWavelets==1.1.1
rembg==1.0.26
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.3
six==1.16.0
starlette==0.14.2
tifffile==2021.4.8
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
waitress==2.0.0
Werkzeug==2.0.0
